Move FormatFileSize to extension method
Hope this is OK. It's because I'm planning to use FileCell.razor in the demo and the fewer lines of code in there, the easier it will be for viewers to make sense of the file. We can change it back if you need the method to be inline in the file for any reason!